backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
alembic
psycopg2-binary
asyncpg
python-dotenv
pytest
pytest-asyncio
httpx
python-multipart
pydantic[email]
pydantic-settings
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
python-jose==3.3.0
pypdf
boto3
